+++ zarrs_codec/src/codec_partial_default.rs
@@ -178,10 +178,6 @@ impl<T: ?Sized> ArrayPartialEncoderTraits
 where
     T: ArrayPartialEncoderTraits,
 {
-    fn into_dyn_decoder(self: Arc<Self>) -> Arc<dyn ArrayPartialDecoderTraits> {
-        self.clone()
-    }
-
     fn erase(&self) -> Result<(), super::CodecError> {
         self.input_output_handle.erase()
     }
@@ -300,10 +296,6 @@ impl<T: ?Sized> ArrayPartialEncoderTraits
 where
     T: BytesPartialEncoderTraits,
 {
-    fn into_dyn_decoder(self: Arc<Self>) -> Arc<dyn ArrayPartialDecoderTraits> {
-        self.clone()
-    }
-
     fn erase(&self) -> Result<(), super::CodecError> {
         self.input_output_handle.erase()
     }
@@ -417,10 +409,6 @@ impl<T: ?Sized> BytesPartialEncoderTraits
 where
     T: BytesPartialEncoderTraits,
 {
-    fn into_dyn_decoder(self: Arc<Self>) -> Arc<dyn BytesPartialDecoderTraits> {
-        self.clone()
-    }
-
     fn erase(&self) -> Result<(), super::CodecError> {
         self.input_output_handle.erase()
     }
@@ -535,10 +523,6 @@ impl<T: ?Sized> AsyncArrayPartialEncoderTraits
 where
     T: AsyncArrayPartialEncoderTraits,
 {
-    fn into_dyn_decoder(self: Arc<Self>) -> Arc<dyn AsyncArrayPartialDecoderTraits> {
-        self.clone()
-    }
-
     async fn erase(&self) -> Result<(), super::CodecError> {
         self.input_output_handle.erase().await
     }
@@ -665,10 +649,6 @@ impl<T: ?Sized> AsyncArrayPartialEncoderTraits
 where
     T: AsyncBytesPartialEncoderTraits,
 {
-    fn into_dyn_decoder(self: Arc<Self>) -> Arc<dyn AsyncArrayPartialDecoderTraits> {
-        self.clone()
-    }
-
     async fn erase(&self) -> Result<(), super::CodecError> {
         self.input_output_handle.erase().await
     }
@@ -789,10 +769,6 @@ impl<T: ?Sized> AsyncBytesPartialEncoderTraits
 where
     T: AsyncBytesPartialEncoderTraits,
 {
-    fn into_dyn_decoder(self: Arc<Self>) -> Arc<dyn AsyncBytesPartialDecoderTraits> {
-        self.clone()
-    }
-
     async fn erase(&self) -> Result<(), super::CodecError> {
         self.input_output_handle.erase().await
     }