+++ zarrs_plugin/src/runtime_plugin.rs
@@ -22,30 +22,32 @@ use crate::{MaybeSend, MaybeSync, PluginCreateError};
 /// );
 /// ```
 #[allow(clippy::type_complexity)]
-pub struct RuntimePlugin<TPlugin, TInput, TError = PluginCreateError>
+pub struct RuntimePlugin<TPlugin, TInput, TError = PluginCreateError, TMatch = str>
 where
     TPlugin: MaybeSend + MaybeSync + 'static,
     TInput: ?Sized + 'static,
+    TMatch: ?Sized,
 {
     /// Tests if the name is a match for this plugin.
-    match_name_fn: Box<dyn Fn(&str) -> bool + Send + Sync + 'static>,
+    match_fn: Box<dyn Fn(&TMatch) -> bool + Send + Sync + 'static>,
     /// Create an implementation of this plugin from input.
     create_fn: Box<dyn Fn(&TInput) -> Result<TPlugin, TError> + Send + Sync + 'static>,
 }
 
-impl<TPlugin, TInput, TError> RuntimePlugin<TPlugin, TInput, TError>
+impl<TPlugin, TInput, TError, TMatch> RuntimePlugin<TPlugin, TInput, TError, TMatch>
 where
     TPlugin: MaybeSend + MaybeSync + 'static,
     TInput: ?Sized + 'static,
+    TMatch: ?Sized,
 {
     /// Create a new runtime plugin for registration.
-    pub fn new<M, C>(match_name_fn: M, create_fn: C) -> Self
+    pub fn new<M, C>(match_fn: M, create_fn: C) -> Self
     where
-        M: Fn(&str) -> bool + Send + Sync + 'static,
+        M: Fn(&TMatch) -> bool + Send + Sync + 'static,
         C: Fn(&TInput) -> Result<TPlugin, TError> + Send + Sync + 'static,
     {
         Self {
-            match_name_fn: Box::new(match_name_fn),
+            match_fn: Box::new(match_fn),
             create_fn: Box::new(create_fn),
         }
     }
@@ -58,10 +60,11 @@ where
         (self.create_fn)(input)
     }
 
-    /// Returns true if this plugin is associated with `name`.
+    /// Returns true if this plugin is associated with `match`.
+    // TODO: Rename to `match` on breaking release
     #[must_use]
-    pub fn match_name(&self, name: &str) -> bool {
-        (self.match_name_fn)(name)
+    pub fn match_name(&self, r#match: &TMatch) -> bool {
+        (self.match_fn)(r#match)
     }
 }
 
@@ -69,32 +72,35 @@ where
 ///
 /// This is the runtime equivalent of [`Plugin2`](crate::Plugin2).
 #[allow(clippy::type_complexity)]
-pub struct RuntimePlugin2<TPlugin, TInput1, TInput2, TError = PluginCreateError>
+pub struct RuntimePlugin2<TPlugin, TInput1, TInput2, TError = PluginCreateError, TMatch = str>
 where
     TPlugin: MaybeSend + MaybeSync + 'static,
     TInput1: ?Sized + 'static,
     TInput2: ?Sized + 'static,
+    TMatch: ?Sized,
 {
     /// Tests if the name is a match for this plugin.
-    match_name_fn: Box<dyn Fn(&str) -> bool + Send + Sync + 'static>,
+    match_fn: Box<dyn Fn(&TMatch) -> bool + Send + Sync + 'static>,
     /// Create an implementation of this plugin from input.
     create_fn: Box<dyn Fn(&TInput1, &TInput2) -> Result<TPlugin, TError> + Send + Sync + 'static>,
 }
 
-impl<TPlugin, TInput1, TInput2, TError> RuntimePlugin2<TPlugin, TInput1, TInput2, TError>
+impl<TPlugin, TInput1, TInput2, TError, TMatch>
+    RuntimePlugin2<TPlugin, TInput1, TInput2, TError, TMatch>
 where
     TPlugin: MaybeSend + MaybeSync + 'static,
     TInput1: ?Sized + 'static,
     TInput2: ?Sized + 'static,
+    TMatch: ?Sized,
 {
     /// Create a new runtime plugin for registration.
-    pub fn new<M, C>(match_name_fn: M, create_fn: C) -> Self
+    pub fn new<M, C>(match_fn: M, create_fn: C) -> Self
     where
-        M: Fn(&str) -> bool + Send + Sync + 'static,
+        M: Fn(&TMatch) -> bool + Send + Sync + 'static,
         C: Fn(&TInput1, &TInput2) -> Result<TPlugin, TError> + Send + Sync + 'static,
     {
         Self {
-            match_name_fn: Box::new(match_name_fn),
+            match_fn: Box::new(match_fn),
             create_fn: Box::new(create_fn),
         }
     }
@@ -107,10 +113,11 @@ where
         (self.create_fn)(input1, input2)
     }
 
-    /// Returns true if this plugin is associated with `name`.
+    /// Returns true if this plugin is associated with `match`.with `match`.
+    // TODO: Rename to `match` on breaking release
     #[must_use]
-    pub fn match_name(&self, name: &str) -> bool {
-        (self.match_name_fn)(name)
+    pub fn match_name(&self, r#match: &TMatch) -> bool {
+        (self.match_fn)(r#match)
     }
 }
 
@@ -121,12 +128,12 @@ where
 mod wasm_impls {
     use super::{RuntimePlugin, RuntimePlugin2};
 
-    unsafe impl<TPlugin: 'static, TInput: ?Sized + 'static, TError: 'static> Send
-        for RuntimePlugin<TPlugin, TInput, TError>
+    unsafe impl<TPlugin: 'static, TInput: ?Sized + 'static, TError: 'static, TMatch: ?Sized> Send
+        for RuntimePlugin<TPlugin, TInput, TError, TMatch>
     {
     }
-    unsafe impl<TPlugin: 'static, TInput: ?Sized + 'static, TError: 'static> Sync
-        for RuntimePlugin<TPlugin, TInput, TError>
+    unsafe impl<TPlugin: 'static, TInput: ?Sized + 'static, TError: 'static, TMatch: ?Sized> Sync
+        for RuntimePlugin<TPlugin, TInput, TError, TMatch>
     {
     }
 
@@ -135,7 +142,8 @@ mod wasm_impls {
         TInput1: ?Sized + 'static,
         TInput2: ?Sized + 'static,
         TError: 'static,
-    > Send for RuntimePlugin2<TPlugin, TInput1, TInput2, TError>
+        TMatch: ?Sized,
+    > Send for RuntimePlugin2<TPlugin, TInput1, TInput2, TError, TMatch>
     {
     }
     unsafe impl<
@@ -143,7 +151,8 @@ mod wasm_impls {
         TInput1: ?Sized + 'static,
         TInput2: ?Sized + 'static,
         TError: 'static,
-    > Sync for RuntimePlugin2<TPlugin, TInput1, TInput2, TError>
+        TMatch: ?Sized,
+    > Sync for RuntimePlugin2<TPlugin, TInput1, TInput2, TError, TMatch>
     {
     }
 }